Check to see if the user is actually a superadmin before returning true
Prior to this commit, if the "laralum.superadmin_bypass_haspermission" config option was set to true, this conditional would *always* return true, regardless of whether or not the user was actually a super admin.  Now it checks to make sure the requesting user is actually a superadmin before returning true.